+++ src/main/java/app/warehouse/system/model/Order.java
@@ -35,5 +35,5 @@ public class Order {
     @OneToMany(mappedBy = "order")
     private Set<Item> orderItems;
     @Column(name = "order_quantity")
-    private Long orderQuantity;
+    private Integer orderQuantity;
 }
